save work, refactor and solved some comments
