+++ crypto/Cargo.toml
@@ -23,7 +23,7 @@ criterion = "0.4"
 iai-callgrind.workspace = true
 rand = "0.8.5"
 rand_chacha = "0.3.1"
-proptest = "1.5.0"
+
 [features]
 default = ["asm", "std"]
 asm = ["sha3/asm"]
