+++ crypto/src/hash/rescue_prime/mds_method.rs
@@ -1,0 +1,10 @@
+#[derive(Clone)]
+#[allow(dead_code)]
+pub enum MdsMethod {
+    /// Use standard matrix multiplication.
+    MatrixMultiplication,
+    /// Use Number Theoretic Transform for multiplication.
+    Ntt,
+    /// Use Karatsuba algorithm for multiplication.
+    Karatsuba,
+}
