+++ crypto/src/hash/rescue_prime/parameters.rs
@@ -5,6 +5,21 @@ pub const ALPHA_INV: u64 = 10540996611094048183;
 // Constants obtained using the paper implementation in Sage
 // https://github.com/ASDiscreteMathematics/rpo/tree/master/reference_implementation
 
+// Constants for the 128-bit security level
+pub const MDS_VECTOR_128: [Fp; 12] = [
+    Fp::const_from_raw(7),
+    Fp::const_from_raw(23),
+    Fp::const_from_raw(8),
+    Fp::const_from_raw(26),
+    Fp::const_from_raw(13),
+    Fp::const_from_raw(10),
+    Fp::const_from_raw(9),
+    Fp::const_from_raw(7),
+    Fp::const_from_raw(6),
+    Fp::const_from_raw(22),
+    Fp::const_from_raw(21),
+    Fp::const_from_raw(8),
+];
 pub const ROUND_CONSTANTS_128: [Fp; 168] = [
     Fp::const_from_raw(5789762306288267392),
     Fp::const_from_raw(6522564764413701783),
@@ -174,6 +189,198 @@ pub const ROUND_CONSTANTS_128: [Fp; 168] = [
     Fp::const_from_raw(9643968136937729763),
     Fp::const_from_raw(3611348709641382851),
     Fp::const_from_raw(18256379591337759196),
+];
+
+pub const MDS_MATRIX_128: [[Fp; 12]; 12] = [
+    [
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+    ],
+    [
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+    ],
+    [
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+    ],
+    [
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+    ],
+    [
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+    ],
+    [
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+    ],
+    [
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+    ],
+    [
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+    ],
+    [
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+    ],
+    [
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+    ],
+    [
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(23),
+    ],
+    [
+        Fp::const_from_raw(23),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(26),
+        Fp::const_from_raw(13),
+        Fp::const_from_raw(10),
+        Fp::const_from_raw(9),
+        Fp::const_from_raw(7),
+        Fp::const_from_raw(6),
+        Fp::const_from_raw(22),
+        Fp::const_from_raw(21),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(7),
+    ],
+];
+
+// Constants for the 160-bit security level
+
+pub const MDS_VECTOR_160: [Fp; 16] = [
+    Fp::const_from_raw(256),
+    Fp::const_from_raw(2),
+    Fp::const_from_raw(1073741824),
+    Fp::const_from_raw(2048),
+    Fp::const_from_raw(16777216),
+    Fp::const_from_raw(128),
+    Fp::const_from_raw(8),
+    Fp::const_from_raw(16),
+    Fp::const_from_raw(524288),
+    Fp::const_from_raw(4194304),
+    Fp::const_from_raw(1),
+    Fp::const_from_raw(268435456),
+    Fp::const_from_raw(1),
+    Fp::const_from_raw(1024),
+    Fp::const_from_raw(2),
+    Fp::const_from_raw(8192),
 ];
 
 pub const ROUND_CONSTANTS_160: [Fp; 224] = [
@@ -403,14 +610,341 @@ pub const ROUND_CONSTANTS_160: [Fp; 224] = [
     Fp::const_from_raw(4582902440098948914),
 ];
 
+pub const MDS_MATRIX_160: [[Fp; 16]; 16] = [
+    [
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+    ],
+    [
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+    ],
+    [
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+    ],
+    [
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+    ],
+    [
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+    ],
+    [
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+    ],
+    [
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+    ],
+    [
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+    ],
+    [
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+    ],
+    [
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+    ],
+    [
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+    ],
+    [
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+    ],
+    [
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+    ],
+    [
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+    ],
+    [
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+        Fp::const_from_raw(2),
+    ],
+    [
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(1073741824),
+        Fp::const_from_raw(2048),
+        Fp::const_from_raw(16777216),
+        Fp::const_from_raw(128),
+        Fp::const_from_raw(8),
+        Fp::const_from_raw(16),
+        Fp::const_from_raw(524288),
+        Fp::const_from_raw(4194304),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(268435456),
+        Fp::const_from_raw(1),
+        Fp::const_from_raw(1024),
+        Fp::const_from_raw(2),
+        Fp::const_from_raw(8192),
+        Fp::const_from_raw(256),
+    ],
+];
+
 pub enum SecurityLevel {
     Sec128,
     Sec160,
 }
 
+// Here if I use [Fp] i get error  abpout two return values
+// if
 pub fn get_round_constants(level: SecurityLevel) -> &'static [Fp] {
     match level {
         SecurityLevel::Sec128 => &ROUND_CONSTANTS_128,
         SecurityLevel::Sec160 => &ROUND_CONSTANTS_160,
+    }
+}
+
+pub enum MdsVector {
+    Mds128([Fp; 12]),
+    Mds160([Fp; 16]),
+}
+
+pub fn get_mds_vector(level: SecurityLevel) -> MdsVector {
+    match level {
+        SecurityLevel::Sec128 => MdsVector::Mds128(MDS_VECTOR_128),
+        SecurityLevel::Sec160 => MdsVector::Mds160(MDS_VECTOR_160),
+    }
+}
+
+#[allow(clippy::large_enum_variant)]
+pub enum MdsMatrix {
+    Mds128([[Fp; 12]; 12]),
+    Mds160([[Fp; 16]; 16]),
+}
+
+pub fn get_mds_matrix(level: SecurityLevel) -> MdsMatrix {
+    match level {
+        SecurityLevel::Sec128 => MdsMatrix::Mds128(MDS_MATRIX_128),
+        SecurityLevel::Sec160 => MdsMatrix::Mds160(MDS_MATRIX_160),
+    }
+}
+
+impl MdsVector {
+    pub fn as_slice(&self) -> &[Fp] {
+        match self {
+            MdsVector::Mds128(ref vec) => vec,
+            MdsVector::Mds160(ref vec) => vec,
+        }
     }
 }
